separated spec files for image box & notification
