+++ cypress/integration/v2/blocks/image-box/desktop-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopStyle } from '.'
+
+describe( 'Image Box Block ( Desktop Style )', registerTests( [ desktopStyle ] ) )
