+++ cypress/integration/v2/blocks/image-box/tablet-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletStyle } from '.'
+
+describe( 'Image Box Block ( Tablet Style )', registerTests( [ tabletStyle ] ) )
