+++ cypress/integration/v2/blocks/notification/mobile-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileStyle } from '.'
+
+describe( 'Notification Block ( Mobile Style )', registerTests( [ mobileStyle ] ) )
